+++ Packages/com.cdm.figma/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.2] - 2023-04-25
+### Changes
+- Added override node option for generating of `Sprite`
+- Added scale factor and min-max texture size parameters to `SpriteGenerateOptions`
+
+### Fixes
+- Fixed copying Figma design content from the Inspector
+
 ## [1.4.0] - 2023-04-23
 ### Changes
 - Added file version option to `FigmaDownloaderAsset`